+++ Build1/zmachine/zmachine/CPUState.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using static zmachine.Machine;
 
 namespace zmachine
 {
