+++ Build1/zmachine/zmachine/Machine.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 
 namespace zmachine
 {
@@ -63,20 +62,6 @@ namespace zmachine
             get => debug;
         }
 
-        // This class stores the chain of routines that have been called, in sequence, and the values in the local variable table.
-        public class RoutineCallState
-        {
-            public ushort[] localVars = new ushort[15];
-            public uint stackFrameAddress = 0;     // Store the stack pointer whenever we call a return
-            public uint returnAddress = 0;         // Store where we need to return to
-            public uint numLocalVars = 0;          // We have an array of localVars, but we don't know how long it is.
-
-            // A stack frame is an index to the routine call state (aka the stack of return addresses for routines already running, and the local variables they carry).
-            // The interpreter should be able to produce the current value and set a value further down the call-stack than the current one, throwing away all others.
-            // We want to be able to call a RoutineCallState for every time a routine is called.
-            // We return values when routine is complete - but routine needs access to local variable table until it returns.
-        } // end RoutineCallState
-
         public uint unpackedAddress(ushort address)
         {
             return (uint)2 * address;
@@ -131,15 +116,6 @@ namespace zmachine
 
             setVar(pc_getByte(), returnVal);                                // Set the return value. Calling a routine is a "store" function, so the next byte contains where to store the result.
         }
-
-        public enum OperandType
-        {
-            Large = 0,
-            Small = 1,
-            Var = 2,
-            Omit = 3
-        };
-
 
         // Find the PC start point in the header file and set PC 
         public void setProgramCounter()
