+++ Build1/zmachine/zmachine/OperandType.cs
@@ -1,0 +1,10 @@
+﻿namespace zmachine
+{
+    public enum OperandType
+    {
+        Large = 0,
+        Small = 1,
+        Var = 2,
+        Omit = 3
+    };
+}
